# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Curator v2 API
  version: 1.10.0
  description: |
    C4 of the frozen contracts (design doc 03). JSON in, JSON out, UTF-8.
    Every route lives under the mount prefix `{base}` (empty by default, `/curation` in
    production; the gateway does not strip it). Write operations are asynchronous: they
    return the resource and its state, progress arrives over SSE or polling.
    Errors use one body, `Error`; `code` is stable for machines, `message` is Chinese for people.
    `precheck_failed` (422, any write that starts work) carries `error.details` shaped as
    `PrecheckDetails`: one entry per pre-start check (D30).
    Responses that point people at a page carry `links` (D22): absolute URLs when
    `server.publicBaseUrl` is configured, otherwise relative paths with `absolute: false`.
    Write endpoints accept only `application/json` - send the header on every write, even without
    a body - and cross-site writes are refused; there is no CORS.
    Ids are opaque strings. Records made since 1.10.0 get `<prefix>-<9 lowercase letters>` (task,
    sub, ds, pf, cred, vb, vm, upl; e.g. `task-kqzmrtbwe`, D45); records made before keep their
    `<prefix>_...` id, and both work everywhere. Ids carry no time: lists sort by creation time.

    1.1.0 (2026-09-21): dataset registry (D36): `/datasets` now lists registered datasets and the
    TOS / cache-bucket listing moved to `/datasets/browse`; `/overview`; starting a task re-checks
...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - basicAuth: []
...
paths:
...
  /vlm-backends:
    get:
      tags: [vlm]
      operationId: listVlmBackends
      summary: List backends with their models
      responses:
        "200":
          description: backends
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: {type: array, items: {$ref: "#/components/schemas/VlmBackend"}}
        default: {$ref: "#/components/responses/Error"}
    post:
      tags: [vlm]
      operationId: createVlmBackend
      summary: Add a backend with its API key; then try GET {endpoint}/models once (D8)
      description: Not listing any models is not an error; the user adds model IDs by hand.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/VlmBackendCreate"}
      responses:
        "201":
          description: saved
          content:
            application/json:
              schema: {$ref: "#/components/schemas/VlmBackend"}
        default: {$ref: "#/components/responses/Error"}
  /vlm-backends/{id}:
    parameters:
      - $ref: "#/components/parameters/PathId"
    put:
      tags: [vlm]
      operationId: updateVlmBackend
      summary: Update; an API key left empty stays unchanged
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/VlmBackendUpdate"}
      responses:
        "200":
          description: updated
          content:
            application/json:
              schema: {$ref: "#/components/schemas/VlmBackend"}
        default: {$ref: "#/components/responses/Error"}
    delete:
      tags: [vlm]
      operationId: deleteVlmBackend
      summary: Delete with its models and key (409 backend_in_use while a non-terminal task uses it)
      description: |
        An unfinished task using it answers 409 in_use. When only finished (or deleted) tasks
        use it, the 409 carries `error.details.confirm_required: true`; repeating the call
        with `confirm=true` deletes it and those tasks lose the reference.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - name: confirm
          in: query
          schema: {type: boolean, default: false}
      responses:
        "204": {description: deleted}
        default: {$ref: "#/components/responses/Error"}
  /vlm-backends/{id}/verify:
    parameters:
      - $ref: "#/components/parameters/PathId"
    post:
      tags: [vlm]
      operationId: verifyVlmBackend
      summary: Probe the endpoint (GET /models, else a minimal chat call)
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: result
          content:
            application/json:
              schema: {$ref: "#/components/schemas/VerifyResult"}
        default: {$ref: "#/components/responses/Error"}
  /vlm-backends/{id}/refresh-models:
    parameters:
      - $ref: "#/components/parameters/PathId"
    post:
      tags: [vlm]
      operationId: refreshVlmModels
      summary: List models again; an empty result means "enter them by hand"
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: listing result
          content:
            application/json:
              schema:
                type: object
                required: [listed, models]
                properties:
                  listed: {type: boolean, description: "false when GET /models is not available"}
                  models: {type: array, items: {$ref: "#/components/schemas/VlmModel"}}
                  note: {type: string}
        default: {$ref: "#/components/responses/Error"}
  /vlm-backends/{id}/models:
    parameters:
      - $ref: "#/components/parameters/PathId"
    post:
      tags: [vlm]
      operationId: addVlmModel
      summary: Add a model ID or inference endpoint ID (ep-...), checked with one minimal call
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/VlmModelCreate"}
      responses:
        "201":
          description: added
          content:
            application/json:
              schema: {$ref: "#/components/schemas/VlmModel"}
        default: {$ref: "#/components/responses/Error"}
  /vlm-backends/{id}/models/{model_id}:
    parameters:
      - $ref: "#/components/parameters/PathId"
      - name: model_id
        in: path
        required: true
        schema: {type: string}
    patch:
      tags: [vlm]
      operationId: updateVlmModel
      summary: Set reasoning effort and parallelism
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/VlmModelPatch"}
      responses:
        "200":
          description: updated
          content:
            application/json:
              schema: {$ref: "#/components/schemas/VlmModel"}
        default: {$ref: "#/components/responses/Error"}
    delete:
      tags: [vlm]
      operationId: deleteVlmModel
      summary: Remove a model (409 backend_in_use while a non-terminal task uses it)
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "204": {description: removed}
        default: {$ref: "#/components/responses/Error"}

  # ---------------------------------------------------------------- datasets
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: HTTP Basic against htpasswd (bcrypt / apr1), realm "Robot Data Curation" (doc 08 §5)
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema: {type: string}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: the same key within 24 hours returns the first response (doc 03 §8)
      schema: {type: string, minLength: 8, maxLength: 128}
...
  responses:
    Error:
      description: |
        400 validation_failed, 401 unauthorized, 404 not_found, 409 task_state_conflict /
        subtask_active / credential_in_use / backend_in_use / name_taken / preflight_expired /
        source_changed / result_changed / idempotency_conflict, 412 precondition_failed,
        422 precheck_failed / confirm_path_mismatch / model_check_failed, 500 internal
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              enum: [validation_failed, unauthorized, not_found, task_state_conflict, subtask_active,
                     credential_in_use, backend_in_use, dataset_in_use, name_taken, preflight_expired, precheck_failed,
                     source_changed, result_changed, confirm_path_mismatch, model_check_failed,
                     idempotency_conflict, precondition_failed, method_not_allowed, internal]
            message: {type: string, description: "Chinese, shown to people as is"}
            details: {type: object}
...
    ReasoningEffort:
      description: Ark's native values; null = the field is not sent (the model's default, and the parity setting)
      enum: [none, minimal, low, medium, high, xhigh, max, null]
    VerifyState:
      enum: [unverified, ok, failed]
    VerifyResult:
      type: object
      additionalProperties: false
      required: [verify_state, last_verified_at]
      properties:
        verify_state: {$ref: "#/components/schemas/VerifyState"}
        last_verified_at: {type: integer}
        error: {type: [string, "null"]}
...
    VlmModel:
      type: object
      additionalProperties: false
      required: [id, model_name, reasoning_effort, max_concurrency, capabilities, source, is_default]
      properties:
        id: {type: string}
        is_default:
          type: boolean
          description: >-
            the model a new task starts with; at most one per owner across every backend. Without
            one the form asks, as before. Deleting the model or its backend leaves none.
        model_name: {type: string, description: "Model ID or inference endpoint ID (ep-...)"}
        reasoning_effort: {$ref: "#/components/schemas/ReasoningEffort"}
        max_concurrency: {type: [integer, "null"], minimum: 1, description: "null = the backend's"}
        capabilities:
          type: object
          additionalProperties: false
          properties:
            vision: {type: [boolean, "null"]}
            reasoning_effort_levels:
              type: array
              description: effective levels to offer (from the site's mapping table; all 7 when unknown)
              items: {enum: [none, minimal, low, medium, high, xhigh, max]}
        source: {enum: [listed, manual]}
    VlmModelCreate:
      type: object
      additionalProperties: false
      required: [model_name]
      properties:
        model_name: {type: string, minLength: 1}
        reasoning_effort: {$ref: "#/components/schemas/ReasoningEffort"}
        max_concurrency: {type: [integer, "null"], minimum: 1}
    VlmModelPatch:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        reasoning_effort: {$ref: "#/components/schemas/ReasoningEffort"}
        max_concurrency: {type: [integer, "null"], minimum: 1}
        is_default:
          type: boolean
          description: >-
            true makes it the one default and clears the flag on every other model of every
            backend; false clears it and leaves none
    VlmBackend:
      type: object
      additionalProperties: false
      required: [id, name, kind, endpoint, max_concurrency, has_api_key, verify_state, models, created_at, updated_at]
      properties:
        id: {type: string}
        name: {type: string}
        kind: {enum: [ark, custom]}
        endpoint: {type: string}
        max_concurrency: {type: integer, minimum: 1, default: 64}
        has_api_key: {type: boolean}
        verify_state: {$ref: "#/components/schemas/VerifyState"}
        last_verified_at: {type: [integer, "null"]}
        last_verify_error: {type: [string, "null"]}
        models: {type: array, items: {$ref: "#/components/schemas/VlmModel"}}
        models_listed: {type: boolean, description: "whether GET /models worked last time"}
        created_at: {type: integer}
        updated_at: {type: integer}
    VlmBackendCreate:
      type: object
      additionalProperties: false
      required: [name, kind, endpoint]
      properties:
        name: {type: string, minLength: 1, maxLength: 64}
        kind: {enum: [ark, custom]}
        endpoint: {type: string, minLength: 1}
        api_key: {type: string, writeOnly: true, description: "required for ark"}
        max_concurrency: {type: integer, minimum: 1}
      if:
        properties: {kind: {const: ark}}
      then:
        required: [api_key]
    VlmBackendUpdate:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name: {type: string, minLength: 1, maxLength: 64}
        endpoint: {type: string}
        api_key: {type: string, writeOnly: true, description: "empty = unchanged"}
        max_concurrency: {type: integer, minimum: 1}

    # ---------- datasets
...